+++ packages/config-registry-controller/src/index.ts
@@ -34,5 +34,8 @@ export type {
   ConfigRegistryApiServiceMethodActions,
 } from './config-registry-api-service/config-registry-api-service-method-action-types.js';
 export type { NetworkFilterOptions } from './config-registry-api-service/filters.js';
-export { ConfigRegistryApiService } from './config-registry-api-service/config-registry-api-service.js';
+export {
+  ConfigRegistryApiService,
+  ConfigRegistryApiEnv,
+} from './config-registry-api-service/config-registry-api-service.js';
 export { filterNetworks } from './config-registry-api-service/filters.js';