+++ packages/config-registry-controller/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add optional `isAutoEnabled?: boolean` property to `RegistryNetworkConfig.config` ([#0000](https://github.com/MetaMask/core/pull/0000))
+- Add optional `isAutoEnabled?: boolean` property to `RegistryNetworkConfig.config` ([#9879](https://github.com/MetaMask/core/pull/9879))
 
 ### Changed